+++ src/App.tsx
@@ -17,7 +17,7 @@ import getCalendar from './utils/calendar-generator'
 
 import i18next from 'i18next'
 import withStyles, { CreateCSSProperties } from '@material-ui/core/styles/withStyles'
-import { createMuiTheme, ThemeProvider, Theme } from '@material-ui/core/styles'
+import { createTheme, ThemeProvider, Theme } from '@material-ui/core/styles'
 
 import Avatar from '@material-ui/core/Avatar'
 import IconButton from '@material-ui/core/IconButton'
@@ -45,7 +45,7 @@ import Typography from '@material-ui/core/Typography'
 import Button from '@material-ui/core/Button'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import { faPaypal } from '@fortawesome/free-brands-svg-icons'
-import Cookies from 'universal-cookie'
+import SavedStateHandler from './utils/saved-state-handler'
 import CardHeader from '@material-ui/core/CardHeader'
 
 import getClasses, { getMinimalClasses } from './utils/shift-scraper'
@@ -54,13 +54,7 @@ import DialogTitle from '@material-ui/core/DialogTitle'
 import DialogActions from '@material-ui/core/DialogActions'
 import DialogContent from '@material-ui/core/DialogContent'
 import Box from '@material-ui/core/Box'
-
-type BuiltCourse = {
-	course: Course,
-	availableShifts: Shift[],
-	selectedShifts: Shift[]
-}
-
+import { APP_STYLES } from './styles/styles'
 
 class App extends React.Component <{
 	classes: CreateCSSProperties
@@ -83,7 +77,7 @@ class App extends React.Component <{
 		darkMode: false,
 		colorPicker: { show: false as boolean, course: undefined as (undefined | Course)  }
 	}
-	cookies = new Cookies()
+	savedStateHandler: SavedStateHandler
 	selectedDegrees: Degree[] = []
 	chosenSchedule: React.RefObject<Schedule>
 	topBar: React.RefObject<TopBar>
@@ -117,38 +111,34 @@ class App extends React.Component <{
 
 		this.theme = this.getTheme(this.state.darkMode)
 
+		this.savedStateHandler = new SavedStateHandler(API.getUrlParams())
+
 		API.setLanguage(this.state.lang)
 	}
 
 	async componentDidMount() {
-		const darkName = this.cookies.get('dark') ?? null
-		if (darkName !== null) {
-			const darkMode = (darkName === 'true') ? true : false
-			if (darkMode !== this.state.darkMode) {
-				this.onChangeDarkMode(darkMode)
-			}
+		const darkMode = this.savedStateHandler.getDarkMode()
+		if (darkMode !== null && darkMode !== this.state.darkMode) {
+			this.onChangeDarkMode(darkMode)
 		}
 
-		const language = this.cookies.get('language') ?? this.state.lang
+		const language = this.savedStateHandler.getLanguage() ?? this.state.lang
 		if (language !== this.state.lang) {
-			this.changeLanguage(language)
+			this.changeLanguage(language, async () => { return })
 		}
 
-		// const currTermId = await defineCurrentTerm()
-		// this.topBar.current?.onSelectedAcademicTerm(currTermId)
-
-		const params = API.getUrlParams()
-		await this.buildState(params.s, params.d)
+		// Build state from cookies or url
+		await this.buildState()
 
 		this.setState({
 			loading: false
 		})
 
 		// Set warning with all notices
-		const isWarned = this.cookies.get('warning')
-		if (isWarned !== 'true') {
+		const isWarned = this.savedStateHandler.getWarning()
+		if (!isWarned) {
 			this.setWarningShiftDegrees()
-			this.cookies.set('warning', 'true', { maxAge: 60*60*24 })
+			this.savedStateHandler.setWarning(true)
 		}
 	}
 
@@ -223,10 +213,7 @@ class App extends React.Component <{
 		})
 
 		this.topBar.current?.setSelectedCourses(currCourses)
-		this.setState({
-			availableShifts,
-			shownShifts
-		})
+		this.setState({ availableShifts, shownShifts })
 	}
 
 	getAllLessons(): Lesson[] {
@@ -238,19 +225,9 @@ class App extends React.Component <{
 	}
 
 	setSelectedShifts(shifts: Shift[]) {
-		this.setState({
-			selectedShifts: shifts
-		})
+		this.setState({ selectedShifts: shifts })
 		this.topBar.current?.setHasSelectedShifts(shifts)
-		if (shifts.length === 0) {
-			this.cookies.remove('s')
-			this.cookies.remove('d')
-		} else {
-			const state = shortenDescriptions(shifts)
-			this.cookies.set('s', state, { maxAge: 60*60*24*31*3 })
-			const degrees = getDegreesAcronyms(shifts)
-			this.cookies.set('d', degrees, { maxAge: 60*60*24*31*3 })
-		}
+		this.savedStateHandler.setShifts(shifts)
 	}
 
 	onSelectedShift(shiftName: string, arr: Shift[]): void {
@@ -297,7 +274,7 @@ class App extends React.Component <{
 				this.showAlert(i18next.t('alert.cleared-schedule'), 'success')
 			}
 
-			this.changeUrl(false)
+			SavedStateHandler.changeUrl(false, [])
 		}
 	}
 
@@ -323,10 +300,7 @@ class App extends React.Component <{
 			availableShifts: this.state.availableShifts
 		})
 
-		this.setState({
-			selectedCampi: campi,
-			shownShifts
-		})
+		this.setState({ selectedCampi: campi, shownShifts })
 	}
 
 	changeShiftTypes(types: string[]): void {
@@ -336,10 +310,7 @@ class App extends React.Component <{
 			availableShifts: this.state.availableShifts
 		})
 
-		this.setState({
-			selectedShiftTypes: types,
-			shownShifts
-		})
+		this.setState({ selectedShiftTypes: types, shownShifts })
 	}
 
 	filterShifts(state: {selectedCampi: string[], selectedShiftTypes: string[], availableShifts: Shift[]}): Shift[] {
@@ -359,9 +330,7 @@ class App extends React.Component <{
 	}
 
 	handleCloseAlert(): void {
-		this.setState({
-			hasAlert: false
-		})
+		this.setState({ hasAlert: false })
 	}
 
 	async getLink(): Promise<void> {
@@ -381,97 +350,41 @@ class App extends React.Component <{
 		this.showAlert(i18next.t('alert.link-obtained'), 'success')
 	}
 
-	async changeUrl(toState: boolean): Promise<void> {
-		const title: string = document.title
-		let path = API.PATH_PREFIX + '/'
-		const state = shortenDescriptions(this.state.selectedShifts)
-		const degrees = getDegreesAcronyms(this.state.selectedShifts)
-		if (state !== '' && toState) {
-			path += `?s=${state}`
-			path += `&d=${degrees}`
-		}
-
-		if (window.history.replaceState) {
-			window.history.replaceState({}, title, path)
-		} else {
-			window.history.pushState({}, title, path)
-		}
-	}
-
-	async buildCourse(description: string[], updates: CourseUpdates): Promise<BuiltCourse> {
-		const course = await API.getCourse(description[0])
-		if (!course) {
-			throw 'Could not build course'
-		}
-
-		if (updates.has(course)) {
-			throw 'Repeated course on URL'
-		}
-
-		updates.toggleCourse(course)
-		const availableShifts = await API.getCourseSchedules(course)
-		if (!availableShifts) {
-			throw 'Could not fetch course schedule'
-		}
-
-		const selectedShiftIds = description.slice(1)
-		const selectedShifts = availableShifts.reduce((acc: Shift[], shift: Shift) => {
-			if (selectedShiftIds.includes(shift.shiftId)) {
-				acc.push(shift)
-				course.addSelectedShift(shift)
-			}
-			return acc
-		}, [] as Shift[])
-		return { course, availableShifts, selectedShifts }
-	}
-
-	async buildState(paramShift: string | undefined, paramDegree: string | undefined): Promise<void> {
-
+	async buildState(forceUpdate = false): Promise<void> {
 		// Build degree
-		paramDegree = paramDegree ?? this.cookies.get('d')
-		if (paramDegree) {
-			try {
-				const degreeAcronyms = paramDegree.split(';')
-				this.topBar.current?.setSelectedDegrees(degreeAcronyms)
-			} catch (err) {
-				console.error(err)
-				// ignored, bad URL/cookie state
-			}
+		try {
+			// Fetch degrees from url params or cookies
+			const degreeAcronyms = this.savedStateHandler.getDegrees(forceUpdate)
+			if (degreeAcronyms) this.topBar.current?.setSelectedDegrees(degreeAcronyms)
+		} catch (err) {
+			console.error(err)
+			// ignored, bad URL/cookie state
 		}
-
 		
 		// Build shifts
-		paramShift = paramShift ?? this.cookies.get('s')
-		if (paramShift) {
-			try {
-				const shifts = paramShift.split(';')
-					.map((shift: string) => shift.split('~'))
-
-				const courseUpdates = new CourseUpdates()
-				const parsedState = await Promise.all(shifts.map(async (description: string[]) => this.buildCourse(description, courseUpdates)))
-				// eslint-disable-next-line
-				const state = parsedState.reduce((acc: any, result: BuiltCourse) => {
-					acc.availableShifts = acc.availableShifts.concat(result.availableShifts)
-					acc.selectedShifts = acc.selectedShifts.concat(result.selectedShifts)
-					return acc
-				}, { availableShifts: [], selectedShifts: [] })
-
-				this.topBar.current?.setSelectedCourses(courseUpdates)
-				this.setState({
-					...state,
-					selectedCourses: courseUpdates,
-					shownShifts: this.filterShifts({
-						selectedCampi: this.state.selectedCampi,
-						selectedShiftTypes: this.state.selectedShiftTypes,
-						availableShifts: state.availableShifts
-					})
-				})
-				this.topBar.current?.setHasSelectedShifts(state.selectedShifts)
-				this.changeUrl(false)
-			} catch (err) {
-				console.error(err)
-				// ignored, bad URL/cookie state
+		try {
+			// Fetch shifts from url params or cookies
+			const shiftState = await this.savedStateHandler.getShifts()
+			if (!shiftState) {
+				return
 			}
+			const [courseUpdates, state] = shiftState
+
+			this.topBar.current?.setSelectedCourses(courseUpdates)
+			this.setState({
+				...state,
+				selectedCourses: courseUpdates,
+				shownShifts: this.filterShifts({
+					selectedCampi: this.state.selectedCampi,
+					selectedShiftTypes: this.state.selectedShiftTypes,
+					availableShifts: state.availableShifts
+				})
+			})
+			this.topBar.current?.setHasSelectedShifts(state.selectedShifts)
+			SavedStateHandler.changeUrl(false, [])
+		} catch (err) {
+			console.error(err)
+			// ignored, bad URL/cookie state
 		}
 	}
 
@@ -482,23 +395,24 @@ class App extends React.Component <{
 		}
 
 		downloadAsImage(this.state.selectedShifts, this.state.darkMode)
-	
 		this.showAlert(i18next.t('alert.schedule-to-image'), 'success')
 	}
 
-	changeLanguage(language: string): void {
-		this.setState({
-			lang: language
-		})
-		i18next.changeLanguage(language).then(() => {
-			i18next.options.lng = language
-		})
-		API.setLanguage(language)
+	async changeLanguage(language: string, afterChange: () => Promise<void>): Promise<void> {
+		if (language !== this.state.lang) {
+			// TODO: Add loader
+			this.setState({loading: true, lang: language })
+			i18next.changeLanguage(language).then(() => i18next.options.lng = language)
+			API.setLanguage(language)
 
-		// Clear shifts?
-		this.clearSelectedShifts(false)
+			// Clear shifts?
+			// this.clearSelectedShifts(false)
+			this.savedStateHandler.setLanguage(language)
 
-		this.cookies.set('language', language, { maxAge: 60*60*24*31*3 })
+			await afterChange()
+			this.buildState(true)
+			this.setState({ loading: false })
+		}
 	}
 
 	onChangeDarkMode(dark: boolean): void {
@@ -506,11 +420,11 @@ class App extends React.Component <{
 		this.setState({
 			darkMode: dark
 		})
-		this.cookies.set('dark', dark, { maxAge: 60*60*24*31*3 })
+		this.savedStateHandler.setDarkMode(dark)
 	}
 
 	getTheme(dark: boolean): Theme {
-		return createMuiTheme({
+		return createTheme({
 			palette: {
 				type: (dark) ? 'dark' : 'light',
 				primary: {
@@ -525,24 +439,11 @@ class App extends React.Component <{
 
 	async getClasses(): Promise<void> {
 		this.setState({loading: true})
-		// const classes = await getClasses(this.state.selectedShifts)
 		const [classesByShift, minimalClasses] = await getMinimalClasses(this.state.selectedShifts, this.selectedDegrees)
-		// const value = Object.entries(classes) //.map(arr => arr.join(': ')).flat() //.join('\r\n')
-		const value = Object.entries(classesByShift)
 
-		this.classesByShift = value
+		this.classesByShift = Object.entries(classesByShift)
 		this.minimalClasses = minimalClasses
 		this.setState({classesDialog: true, loading: false})
-
-		// TODO: download to a file
-		// const el = document.createElement('a')
-		// const file = new Blob([classesByShift], {type: 'text/plain;charset=utf-8'})
-		// el.href = URL.createObjectURL(file)
-		// el.download = 'ist-turmas.txt'
-		// document.body.appendChild(el)
-		// el.click()
-		// document.body.removeChild(el)
-		// this.showAlert(i18next.t('alert.classes-file'), 'success')
 	}
 
 	setWarningShiftDegrees(): void {
@@ -583,7 +484,6 @@ class App extends React.Component <{
 				},
 			}
 		}))(ToggleButtonGroup)
-
 
 		return (
 			<ThemeProvider theme={this.theme}>
@@ -850,84 +750,4 @@ class App extends React.Component <{
 	}
 }
 
-const cssVariables = {
-	blur: '5px',
-	brightness: 1
-}
-
-// eslint-disable-next-line
-const styles = (theme: any) => ({
-	backdrop: {
-		zIndex: theme.zIndex.drawer + 1,
-		color: '#fff',
-		background: 'rgba(0,0,0,0.85)',
-	},
-	checklistSelected: {
-		// color: theme.palette.text.primary
-	},
-	checklistUnselected: {
-		// color: theme.palette.text.hint
-		opacity: 0.8
-	},
-	paper: {
-		display: 'flex',
-		flexWrap: 'wrap' as const,
-		// border: `1px solid ${theme.palette.divider}`,
-	},
-	divider: {
-		margin: theme.spacing(1, 0.5),
-	},
-	toggleGroup: {
-		flexWrap: 'wrap' as const
-	},
-	card: {
-		margin: '1% 1% 2% 1%'
-	},
-	cardTitle: {
-		padding: '8px 16px 2px 16px'
-	},
-	cardContent: {
-		paddingTop: '4px',
-		paddingBottom: '0px'
-	},
-	contentCopyable: {
-		userSelect: 'text' as const,
-		webkitUserSelect: 'text' as const,
-		oUserSelected: 'text' as const
-	},
-	footer: {
-		bottom: '0px',
-		top: 'auto',
-	},
-	grow: {
-		flexGrow: 1,
-	},
-	centered: {
-		margin: 'auto'
-	},
-	body: {
-		height: '100%',
-		'&::before': {
-			content: '""',
-			position: 'fixed',
-			top: '-5%',
-			left: '-5%',
-			right: 0,
-			zIndex: -1,
-
-			display: 'block',
-			backgroundImage: `url(${process.env.PUBLIC_URL}/img/background.jpg)`,
-			backgroundSize: 'cover',
-			width: '110%',
-			height: '110%',
-
-			webkitFilter: `blur(${cssVariables.blur}) brightness(${cssVariables.brightness})`,
-			mozFilter: `blur(${cssVariables.blur}) brightness(${cssVariables.brightness})`,
-			oFilter: `blur(${cssVariables.blur}) brightness(${cssVariables.brightness})`,
-			msFilter: `blur(${cssVariables.blur}) brightness(${cssVariables.brightness})`,
-			filter: `blur(${cssVariables.blur}) brightness(${cssVariables.brightness})`,
-		}
-	}
-})
-
-export default withStyles(styles)(App)
+export default withStyles(APP_STYLES)(App)